+++ packages/redux/src/actions/content.ts
@@ -16,6 +16,7 @@
 
 import { Descriptor, Item, NavigationItem } from '@craftercms/models';
 import { createAction } from "@reduxjs/toolkit";
+import { internal_getNav, internal_getNavBreadcrumb } from "./content_internal";
 
 export const getItem = /*#__PURE__*/ createAction<string>('GET_ITEM');
 
@@ -33,10 +34,48 @@ export const getTree = /*#__PURE__*/ createAction<{ url: string, depth?: Number 
 
 export const getTreeComplete = /*#__PURE__*/ createAction<{ url: string, tree?: Item }>('GET_TREE_COMPLETE');
 
-export const getNav = /*#__PURE__*/ createAction<{ url: string, depth?: Number, currentPageUrl?: string }>('GET_NAV');
+interface GetNavAction {
+  type: string;
+  payload: {
+    url: string;
+    depth?: Number;
+    currentPageUrl?: string;
+  }
+}
+
+export function getNav(url: string): GetNavAction;
+export function getNav(url: string, depth: Number): GetNavAction;
+export function getNav(url: string, depth: Number, currentPageUrl: string): GetNavAction;
+export function getNav(payload: { url: string; depth?: Number; currentPageUrl?: string }): GetNavAction;
+export function getNav(payloadOrUrl: string | { url: string; depth?: Number; currentPageUrl?: string }, depth: Number = 1, currentPageUrl: string = ''): GetNavAction {
+  if (typeof payloadOrUrl === 'string') {
+    console.warn('Warning: This signature is deprecated, adjust to use the object signature instead ({ url, depth?, currentPageUrl? }).');
+    return internal_getNav({ url: payloadOrUrl, depth, currentPageUrl });
+  } else {
+    return internal_getNav(payloadOrUrl);
+  }
+}
 
 export const getNavComplete = /*#__PURE__*/ createAction<{ url: string, nav?: NavigationItem }>('GET_NAV_COMPLETE');
 
-export const getNavBreadcrumb = /*#__PURE__*/ createAction<{ url: string, root?: string }>('GET_NAV_BREADCRUMB');
+export interface GetNavBreadcrumbAction {
+  type: string;
+  payload: {
+    url: string;
+    root?: string;
+  }
+}
+
+export function getNavBreadcrumb(url: string): GetNavBreadcrumbAction;
+export function getNavBreadcrumb(url: string, root: string): GetNavBreadcrumbAction;
+export function getNavBreadcrumb(payload: { url: string, root?: string }): GetNavBreadcrumbAction;
+export function getNavBreadcrumb(payloadOrUrl: string | { url: string, root?: string }, root: string = ''): GetNavBreadcrumbAction {
+  if (typeof payloadOrUrl === 'string') {
+    console.warn('Warning: This signature is deprecated, adjust to use the object signature instead ({ url, root? }).');
+    return internal_getNavBreadcrumb({ url: payloadOrUrl, root });
+  } else {
+    return internal_getNavBreadcrumb(payloadOrUrl);
+  }
+}
 
 export const getNavBreadcrumbComplete = /*#__PURE__*/ createAction<{ url: string, breadcrumb?: Array<NavigationItem> }>('GET_NAV_BREADCRUMB_COMPLETE');