+++ packages/redux/src/actions/content_internal.ts
@@ -14,23 +14,8 @@
  * along with this program. If not, see http://www.gnu.org/licenses/.
  */
 
-export interface ContentInstanceSystemProps {
-  id: string;
-  path: string | null;
-  label: string; // Internal name
-  dateCreated: string;
-  dateModified: string;
-  contentTypeId: string;
-  sourceMap?: Record<string, string>; // { fieldId: path }
-  disabled: boolean;
-  orderInNav?: number;
-  placeInNav?: boolean;
-}
+import { createAction } from "@reduxjs/toolkit";
 
-export interface ContentInstanceBase {
-  craftercms: ContentInstanceSystemProps;
-}
+export const internal_getNav = /*#__PURE__*/ createAction<{ url: string, depth?: Number, currentPageUrl?: string }>('GET_NAV');
 
-export type ContentInstance<T extends Record<string, any> = Record<string, any>> = T & ContentInstanceBase;
-
-export default ContentInstance;
+export const internal_getNavBreadcrumb = /*#__PURE__*/ createAction<{ url: string, root?: string }>('GET_NAV_BREADCRUMB');